pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "travel-agent"
version = "0.1.0"
description = "Travel agent demo using the Strands Graph pattern."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents[ollama]",
    "fastapi",
    "uvicorn",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "jsonschema",
]

[project.scripts]
travel-agent = "travel_agent.main:main"
travel-agent-api = "travel_agent.server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: requires local Ollama and is skipped by default"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
